requirements.txt
# Gmail to Google Group Migration Tool Requirements

# Google API libraries
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
google-auth>=2.20.0

# Configuration and utilities
PyYAML>=6.0
python-dateutil>=2.8.2

# HTTP and networking
httplib2>=0.20.4
requests>=2.28.0

# Logging and monitoring
colorlog>=6.7.0

# Data handling
email-validator>=2.0.0






